Added AdminAuthorizationController and corresponding unit tests
Created an AdminAuthorizationController that interacts with the AdminModel to handle the authorization of admins by managing the login and registration workflows. This class contains method to display the login view to allow the admin to log in, and a method to display the admin dashboard. It also contains method to handle the registration of a new admin to the quiz system, and a method to handle the admin login process to the admin dashboard. Created unit tests to test the admin authorization, and to test the register of an admin. Created unit tests to test the initialization of the view, and created integrity tests for the AdminAuthorizationController class. I also fixed LoginView and RegisterView and its tests.